Make judging season/year more specific
Instead of referencing the season as just `twenty_one` for example, it
will be more specific and accurate now as `twenty_twenty_one`.

Refs: #3292
