+++ app/services/judging/twenty_twenty_two/beginner_questions.rb
@@ -1,5 +1,5 @@
 module Judging
-  module TwentyTwo
+  module TwentyTwentyTwo
     class BeginnerQuestions
       def call
         [
