+++ app/services/judging/twenty_twenty_two/junior_questions.rb
@@ -1,6 +1,6 @@
 module Judging
-  module TwentyOne
-    class SeniorQuestions
+  module TwentyTwentyTwo
+    class JuniorQuestions
       def call
         [
           Question.new(
@@ -9,9 +9,9 @@ module Judging
             field: :ideation_1,
             worth: 5,
             text: %(
-              Team has chosen an important problem and uses statistics,
-              facts and personal stories to demonstrate its impact on them
-              and/or their community.
+              Did we demonstrate the problem we chose is important, and use
+              statistics, facts and personal stories to show its impact on
+              us and our community or the world?
             )
           ),
 
@@ -21,8 +21,10 @@ module Judging
             field: :ideation_2,
             worth: 5,
             text: %(
-              To ensure positive impact of solution, team has completed
-              user research and adapted their idea based on community feedback.
+              Will our app help solve a
+              <a href="https://sdgs.un.org/goals" class="font-weight--bold" target="_blank">UN SDG</a>
+              and positively impact direct and
+              indirect users? Did we conduct solid user research?
             )
           ),
 
@@ -31,7 +33,11 @@ module Judging
             section: "ideation",
             field: :ideation_3,
             worth: 5,
-            text: "Team presents a fundamentally new solution or use of mobile app technology."
+            text: %(
+              Do you agree that our app is innovative (a fundamentally
+              new solution/use of mobile app technology or an innovative
+              application of an existing structure to a new situation)?
+            )
           ),
 
           Question.new(
@@ -39,7 +45,10 @@ module Judging
             section: "ideation",
             field: :ideation_4,
             worth: 5,
-            text: "Team's app or idea was improved or changed in response to competitor research."
+            text: %(
+              Do you see evidence that our app or idea was improved or
+              changed in response to competitor research?
+            )
           ),
 
           Question.new(
@@ -47,7 +56,10 @@ module Judging
             section: "technical",
             field: :technical_1,
             worth: 5,
-            text: "App is fully functional and has no noticeable bugs."
+            text: %(
+              Do we show that our app is fully functional in the Demo Video
+              or by launching it in an app store?
+            )
           ),
 
           Question.new(
@@ -56,8 +68,9 @@ module Judging
             field: :technical_2,
             worth: 5,
             text: %(
-              Team provides examples of how they developed app for a target audience
-              and shows evidence of testing and refinement.
+              Do we demonstrate how we developed our app for our target
+              audience, tested it with them, and made sure it was easy to
+              use?
             )
           ),
 
@@ -66,7 +79,7 @@ module Judging
             section: "technical",
             field: :technical_3,
             worth: 5,
-            text: "All team members appear to have gained technical knowledge and contributed to coding."
+            text: "Are you able to see what our team learned about coding?"
           ),
 
           Question.new(
@@ -75,9 +88,9 @@ module Judging
             field: :technical_4,
             worth: 5,
             text: %(
-              Code includes advanced functions such as using a database with APIs
-              and/or app uses more than 1 sensor, phone function, or different
-              technology (like AI, VR or hardware).
+              Does our code include advanced functions such as using a
+              database with APIs and/or using more than 1 sensor, phone
+              function, or different technology (like AI, VR or hardware)?
             )
           ),
 
@@ -87,8 +100,8 @@ module Judging
             field: :pitch_1,
             worth: 5,
             text: %(
-              The pitch makes me believe that this is an urgent problem and that the
-              team has an effective solution.
+              Does our pitch video convey the urgency of our problem and
+              solution in a creative and engaging way?
             )
           ),
 
@@ -98,8 +111,8 @@ module Judging
             field: :pitch_2,
             worth: 5,
             text: %(
-              The pitch makes me believe that the team learned from challenges and
-              each girl persevered to learn new skills.
+              Do you see evidence of our team sharing our journey, including
+              challenges we encountered and how we grew?
             )
           ),
 
@@ -109,8 +122,9 @@ module Judging
             field: :entrepreneurship_1,
             worth: 5,
             text: %(
-              Team has clear goals and plan to reach target users.
-              They've integrated feedback from initial marketing attempts into this plan.
+              How clearly has our team defined our goals, planned to reach
+              target users, and integrated feedback from initial marketing
+              attempts into our plan?
             )
           ),
 
@@ -120,8 +134,8 @@ module Judging
             field: :entrepreneurship_2,
             worth: 5,
             text: %(
-              Team has a financial plan supported by budgets and research for
-              starting and sustaining the business.
+              How realistic and thorough do you find our financial plan to
+              be, and is it supported by budgets and research?
             )
           ),
 
@@ -131,8 +145,9 @@ module Judging
             field: :entrepreneurship_3,
             worth: 5,
             text: %(
-              Business plan includes logical company, product or service descriptions,
-              market analysis, and graphics.
+              How cohesive and realistic do find our business plan? Does
+              it include logical company, product or service descriptions,
+              market analysis, and graphics?
             )
           ),
 
@@ -141,7 +156,7 @@ module Judging
             section: "entrepreneurship",
             field: :entrepreneurship_4,
             worth: 5,
-            text: "Branding is clear and amplifies the team’s purpose."
+            text: "Do you believe our branding is clear and amplifies our team's purpose?"
           ),
 
           Question.new(
@@ -149,7 +164,7 @@ module Judging
             section: "overall",
             field: :overall_1,
             worth: 5,
-            text: "I believe this team will continue working to make their ideas a reality."
+            text: "Are our team's proposed plan and goals to continue working on our app realistic?"
           ),
 
           Question.new(
@@ -157,10 +172,7 @@ module Judging
             section: "overall",
             field: :overall_2,
             worth: 5,
-            text: %(
-              This solution will positively impact our world!
-              The idea is well thought out and the app is developed.
-            )
+            text: "How well do you think our solution is thought out and can succeed?"
           )
         ]
       end
