+++ spec/services/judge_questions_for_submission_score_spec.rb
@@ -30,7 +30,7 @@ describe JudgeQuestionsForSubmissionScore do
     let(:submission_score_for_ideation_1) { 5 }
 
     before do
-      allow(Judging::TwentyTwo::BeginnerQuestions).to receive_message_chain(:new, :call).and_return(questions)
+      allow(Judging::TwentyTwentyTwo::BeginnerQuestions).to receive_message_chain(:new, :call).and_return(questions)
       allow(submission_score).to receive(:ideation_1).and_return(submission_score_for_ideation_1)
     end
 
@@ -45,9 +45,9 @@ describe JudgeQuestionsForSubmissionScore do
 
   describe "getting the correct questions for a given season and division" do
     before do
-      allow(Judging::TwentyTwo::SeniorQuestions).to receive_message_chain(:new, :call).and_return([])
-      allow(Judging::TwentyTwo::JuniorQuestions).to receive_message_chain(:new, :call).and_return([])
-      allow(Judging::TwentyTwo::BeginnerQuestions).to receive_message_chain(:new, :call).and_return([])
+      allow(Judging::TwentyTwentyTwo::SeniorQuestions).to receive_message_chain(:new, :call).and_return([])
+      allow(Judging::TwentyTwentyTwo::JuniorQuestions).to receive_message_chain(:new, :call).and_return([])
+      allow(Judging::TwentyTwentyTwo::BeginnerQuestions).to receive_message_chain(:new, :call).and_return([])
     end
 
     context "when it's the 2022 season" do
@@ -57,7 +57,7 @@ describe JudgeQuestionsForSubmissionScore do
         let(:division) { "senior" }
 
         it "calls the appropriate class to get the 2022 senior questions" do
-          expect(Judging::TwentyTwo::SeniorQuestions).to receive_message_chain(:new, :call)
+          expect(Judging::TwentyTwentyTwo::SeniorQuestions).to receive_message_chain(:new, :call)
 
           judge_questions_for_submission_score
         end
@@ -67,7 +67,7 @@ describe JudgeQuestionsForSubmissionScore do
         let(:division) { "junior" }
 
         it "calls the appropriate class to get the 2022 junior questions" do
-          expect(Judging::TwentyTwo::JuniorQuestions).to receive_message_chain(:new, :call)
+          expect(Judging::TwentyTwentyTwo::JuniorQuestions).to receive_message_chain(:new, :call)
 
           judge_questions_for_submission_score
         end
@@ -77,7 +77,7 @@ describe JudgeQuestionsForSubmissionScore do
         let(:division) { "beginner" }
 
         it "calls the appropriate class to get the 2022 beginner questions" do
-          expect(Judging::TwentyTwo::BeginnerQuestions).to receive_message_chain(:new, :call)
+          expect(Judging::TwentyTwentyTwo::BeginnerQuestions).to receive_message_chain(:new, :call)
 
           judge_questions_for_submission_score
         end
